Add changes
apt-get will be used over apt. This is mainly to
respect that apt-get is meant for scripts,
where as apt is not.

All the system dependencies will be passed in
to apt-get once. The idea being I should be able
to make sense of the output from apt-get
if things go janky.

Comment tweaks and stdout output from echos
has also been changed.