Split .env into diff files
.env is now split into: .test_env, .test_env_status_codes,
.test_env_local, and .test_env_secrets.
This is allow more additions into the idea of the env
used for testing without it becoming a giant clump.

install has now been changed to accommodate the
generation of .test_env_local, and .test_env_secrets.
These files will change for each install, that, and secrets generally shouldn't be in cvs.

That, and putting env vars that are
meant to be secrets in a file that is cvs is not a great idea either.