Discard clear command in install
This is because clear requires the TERM env var
and during image construction, this var is not set.

While a simple check could either allow or disallow this command,
I don't believe this really added a whole lot anyways (e.g. benefits,
just looking cool). That, and the shell session will need to be closed
upon installation.