Refactor jenkins's files
Will add, as of currently, any changes made to shellfuncs file(s)
such as 'install' will only take effect when the entire repo
is reinstalled. As of currently the goal is to have most of the
customization set in the shell provisioner. So this maybe ok for now.

Otherwise, the build script for shellfuncs will have to take in
account for installing the repo itself. Which may or may not be a lot.