Delete bogus code in test driver, add to heredoc
The git commands in the first heredoc have been modified.
Git pull has been added incase the repo on the VM is out of
date and could benefit from pulling those changes.

Git checkout has been silenced in favor of a custom echo.
I felt like the output was a bit to much and unrelated to
the checkout itself (more detailing about being in a detached
HEAD state).