Fix install_tortoisehg w/being ran, update driver
The install tortoisehg script currently would error out if you
tried to run it more than once. This was probably ok to begin
with but now for good testing, a workaround has been placed.

Test driver also has a initial implementation in to tell jenkins
whether or not all tests have passed or not.